Ajout de commentaire XML
Ajout de commentaire XML et léger changement dans les script Inventory_test et Items et ItemsSlots
