+++ Kamel Quest/Assets/Script/Battle.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb7ea80756d8a804e9c31bcd845ae560
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
