+++ Kamel Quest/Assets/Script/Dialogu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bbf8cbe5dfa01df47bb274f51daf2583
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
