+++ Kamel Quest/Assets/Script/Dialogu/Dialogue.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Dialogue : MonoBehaviour
+{
+    public List<string> SentencesList;
+}
