+++ Kamel Quest/Assets/Script/Dialogu/ShowsBubbleText.cs.meta
