+++ Kamel Quest/Assets/Script/Dialogu/ShowsText.cs.meta
