--- Kamel Quest/Assets/Script/Dialogue.cs
@@ -1,8 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Dialogue : MonoBehaviour
-{
-    public List<string> SentencesList;
-}
