+++ Kamel Quest/Assets/Script/Interface.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f2e365f5ee1d85479e6dab37b51b9d8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
