+++ Kamel Quest/Assets/Script/Interface/Pause.cs
