+++ Kamel Quest/Assets/Script/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c9de3246e78779c48baedd782c559e92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
