+++ Kamel Quest/Assets/Script/Quest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f63bcbda3f82ae545bba03f7442d9545
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
