+++ Kamel Quest/Assets/Script/Quest/QuestGoal.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class QuestGoal
+{
+    public QuestType type;
+    public bool succeded;
+
+}
+
