--- Kamel Quest/Assets/Script/QuestGoal.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class QuestGoal
-{
-    public QuestType type;
-    public bool succeded;
-
-}
-
