+++ Kamel Quest/ProjectSettings/InputManager.asset
@@ -326,11 +326,11 @@ InputManager:
     axis: 0
     joyNum: 0
   - serializedVersion: 3
-    m_Name: Cancel
+    m_Name: Test2
     descriptiveName: 
     descriptiveNegativeName: 
     negativeButton: 
-    positiveButton: escape
+    positiveButton: t
     altNegativeButton: 
     altPositiveButton: joystick button 1
     gravity: 1000
@@ -342,11 +342,11 @@ InputManager:
     axis: 0
     joyNum: 0
   - serializedVersion: 3
-    m_Name: Cancel
+    m_Name: Test3
     descriptiveName: 
     descriptiveNegativeName: 
     negativeButton: 
-    positiveButton: escape
+    positiveButton: y
     altNegativeButton: 
     altPositiveButton: joystick button 1
     gravity: 1000
