+++ Kamel Quest/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.2.21f1
-m_EditorVersionWithRevision: 2019.2.21f1 (9d528d026557)
+m_EditorVersion: 2019.2.19f1
+m_EditorVersionWithRevision: 2019.2.19f1 (929ab4d01772)
